project 2/requirements.txt
tqdm
pytest
scipy
torchvision
webcolors
matplotlib
torch
numpy
opencv_python
Pillow
skimage
